javított double és focus kezelés

+++ app/src/main/java/com/fusetech/mobilleltarkotlin/dataItems/RaktarAdat.kt
@@ -1,6 +1,4 @@
 package com.fusetech.mobilleltarkotlin.dataItems
 
-import java.math.BigDecimal
-
-data class RaktarAdat(val cikkszam: String, val megnevezes1: String?, val megnevezes2: String?, val mennyiseg: BigDecimal, val megjegyzes: String?, val bizszam: Int){
+data class RaktarAdat(val cikkszam: String, val megnevezes1: String?, val megnevezes2: String?, val mennyiseg: Double, val megjegyzes: String?, val bizszam: Int){
 }
